プレイヤーの当たり判定、サイズ変更・ゲームパッド対応

+++ competition/Objects/Character/Player/Player.h
@@ -38,7 +38,7 @@ public:
 	void Movement(float);
 	void Animation();
 
-	void AddExperience(int exp);
+	void AddExperience(float exp);
 	void LevelUp();
 	void SubtractLife();
 	void AddShield(int shield);
